Fixed some bugs

+++ src/Controller.hpp
@@ -64,7 +64,7 @@ private:
 	
 	/***** Action-space search and learning-related constants *****/
 	const int learningMode = 1;							// Which weight adjustment algorithm to use. Supported modes: 0 (no learning), 1 (racist gradient descent).
-	const bool useSig = false;							// Whether to use sigmoid functions or not.
+	const bool useSig = false;							// Whether to use sigmoid functions in network evaluation.
 	const float defaultStepSize = 0.001; 				// Learning rate; multiplies learned outcome's impact on network node weights
 	const float actionDepth = 5; 						// How many variations of acceleration/turning values to test. Primary performance impact in network eval. Up to ~200 should be manageable.
 	const float discountFactor = 0.1; 					// [0.0f, 1.0f); importance of "later" vs "now", with higher values increasing the impact of "now"
@@ -77,8 +77,8 @@ private:
 	
 	/***** Reward function coefficients, see reward in Controller::takeStep() *****/
 	const float timeToFitnessMultiplier = 1.2;			// Unused. Fitness function balancing multiplier.
-	const float wallPenalty = -10000.0;					// Reward penalty for hitting a wall.
-	const float prevVelocityCoefficient = 0.9;			// Reward multiplier for increasing speed vs going fast. Higher value means that increasing car speed is good. Only [0.0f, 1.0f] are sensible.
+	const long double wallPenalty = -10000.0;			// Reward penalty for hitting a wall.
+	const long double prevVelocityCoefficient = 0.9;	// Reward multiplier for increasing speed vs going fast. Higher value means that increasing car speed is good. Only [0.0f, 1.0f] are sensible.
 
 	/***** Controller variables *****/
 	b2World* m_world;
